+++ tasks/terraform/src/main/java/com/walmartlabs/concord/plugins/terraform/actions/Action.java
@@ -33,12 +33,8 @@ import java.io.IOException;
 import java.io.OutputStream;
 import java.nio.file.Files;
 import java.nio.file.Path;
-import java.nio.file.Paths;
 import java.nio.file.StandardOpenOption;
-import java.util.Collections;
-import java.util.List;
 import java.util.Map;
-import java.util.stream.Collectors;
 
 public abstract class Action {
 
@@ -55,14 +51,6 @@ public abstract class Action {
         }
 
         return p;
-    }
-
-    protected static List<Path> createUserSuppliedVarFiles(Path dir, List<String> userSuppliedVarFiles) {
-        if (userSuppliedVarFiles == null) {
-            return Collections.emptyList();
-        }
-
-        return userSuppliedVarFiles.stream().map(f -> dir.resolve(f)).collect(Collectors.toList());
     }
 
     protected static void init(Context ctx, Path workDir, Path dir, boolean silent, Map<String, String> env, Terraform terraform, Backend backend) throws Exception {